feat: migrate an api route to ts functions
